refactor : paymentMethodController
