+++ lib/app/pages/manage_method/binding.dart
@@ -1,11 +1,13 @@
 import 'package:dimipay/app/data/modules/payment_method/controller.dart';
 import 'package:dimipay/app/data/modules/payment_method/repository.dart';
 import 'package:dimipay/app/data/provider/api.dart';
+import 'package:dimipay/app/pages/manage_method/controller.dart';
 import 'package:get/get.dart';
 
 class ManageMethodPageBinding implements Bindings {
   @override
   void dependencies() {
     Get.lazyPut<PaymentMethodController>(() => PaymentMethodController(PaymentMethodRepository(ApiProvider())));
+    Get.lazyPut<ManageMethodPageController>(() => ManageMethodPageController());
   }
 }
