+++ lib/app/pages/manage_method/controller.dart
@@ -1,0 +1,14 @@
+import 'package:dimipay/app/data/modules/payment_method/controller.dart';
+import 'package:get/get.dart';
+
+class ManageMethodPageController extends GetxController {
+  final PaymentMethodController paymentMethodController = Get.find<PaymentMethodController>();
+
+  @override
+  void onInit() {
+    if (paymentMethodController.paymentMethods == null) {
+      paymentMethodController.fetchPaymentMethods();
+    }
+    super.onInit();
+  }
+}
