+++ lib/app/pages/user/controller.dart
@@ -1,14 +1,19 @@
+import 'package:dimipay/app/data/modules/payment_method/controller.dart';
 import 'package:dimipay/app/data/modules/user/controller.dart';
 import 'package:flutter/services.dart';
 import 'package:get/get.dart';
 
 class UserPageController extends GetxController {
   UserController userController = Get.find<UserController>();
+  PaymentMethodController paymentMethodController = Get.find<PaymentMethodController>();
 
   @override
   void onInit() {
     if (userController.user == null) {
       userController.fetchUser();
+    }
+    if (paymentMethodController.paymentMethods == null) {
+      paymentMethodController.fetchPaymentMethods();
     }
     super.onInit();
   }
